requirements.txt
feedparser
youtube-transcript-api
requests
google-generativeai
